requirements.txt
pip==20.2.2
beautifulsoup4~=4.9.1
mechanize~=0.4.5
pymongo~=3.11.0
dnspython~=2.0.0
python_dotenv~=0.14.0